kernel_patch_verify: switch to using olddefconfig
When a patch results in a change in kernel configuration
such that a config variable needs to have a new value,
'make oldconfig' will wait for user to provide input for
that new config.

This makes the 'kernel_patch_verify' appear hung at the
defconfig step (since command output is also redirected).
The build proceeds further only when user presses return
key, for example.

Fix this by switching to olddefconfig instead, which is
non-interactive and starts using the default answers to
new config questions.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>