kernel_patch_verify: Drop parallel builds for yaml checks
YAML checks do a bunch of cleanups of intermediate files, so there is no
real point in attempting parallel checks since we endup stepping on our
own toes. In addition, we also need to introduce additional checks when
we introduce yaml files, so lets make sure we do it serially.
bindings_check for a specific yaml file introduction is'nt usually too
expensive time wise, so lets setup the system for to start expanding our
test coverage.

Signed-off-by: Nishanth Menon <nm@ti.com>