Dockerfile: Bundle the LLVM public key
We can fail to download this sometimes especially behind a corporate
firewall that forces wget to use a proxy. Just bundle this for now.

This also allows us to skip the second apt update and key install
step which throws a warning on Debian 11.

The key can be downloaded with
$ wget https://apt.llvm.org/llvm-snapshot.gpg.key

Signed-off-by: Andrew Davis <afd@ti.com>