build-env.sh: Be consistent with install directories
All other tools are installed into /usr/local except the ARM
cross-compilers which go to /opt. There is no reason for this
so lets install everything to /usr/local.

In the kps/kpv scripts we add these to the PATH in addition
to the /opt version, this way if you do passthrough an /opt
version it will still work.

Signed-off-by: Andrew Davis <afd@ti.com>