Dockerfile: Simplify user ID passthrough
Docker containers by default use the root user. This can cause issues
on some systems and leads to generated files passed through to the host
system to have root permissions. To avoid this we would like to run
our container as a user with our same UID/GID.

Previously we did that by creating the container image with a user
with our UID. The issue is this will only match for the user who
created the image. When we upload this and another user download
the image the built in UID might not match their own. To workaround
that we would then in kps/kpv overwrite the created user info
by making temp /etc/passwd and /etc/group files and passing these
into the container.

This undid the original user creation we did in the dockerfile, but
also creates extra temp files on the host.

Lets instead keep root as the default user in our container image
and overwrite this in the scripts at runtime by making a temporary
container with a new user with our UID/GID.

Signed-off-by: Andrew Davis <afd@ti.com>