build-env.sh/kpv: Fix issues with smatch
https://repo.or.cz/smatch.git was derived out of sparse.

This implies that the output binaries share the same name :(.
Unfortunately, this also means that when we try to create an
environment with sparse AND smatch installed, we end up overwriting
one with the other as is the case right now (smatch is overwritten by
sparse). These two tools have benefits, so lets install smatch in a
completely different non-standard folder, and use the recently created
extension of kernel_patch_verify to provide a custom checker script to
send the data here.

Signed-off-by: Nishanth Menon <nm@ti.com>