kernel_patch_verify: ytest_dtbs_check: DO NOT delete the yamls
When we build check against all other dtbs for the given new yaml
change, we have a bug: if the yaml file was modified, yaml
bindings_check will also generate example.dtb. but we go and delete all
yaml files in the folder where example.dtb is located, which in turns
out to be the very same yaml we are wanting to test against.

So, just delete the dtbs.

Reported-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>