all: replace all instances of $* with "$@"
Using "$@" will prevent unnecessary splitting of quoted parameters. For
instance, if the following parameter was passed previously: "-a 'hello
world'" it would be split into "-a" "hello" "world" which could break
things later in processing.

The kernel_patch_verify script will have to be reworked to deal with
this independently as a lot of functions inherently rely on splitting
for handling arrays.

Signed-off-by: Randolph Sapp <res.sapp@gmail.com>