Dockerfile: remove duplicate sections
We don't need to build 2 different images in the middle of this. There
is substantial overlab between the build tools required to boostrap our
tools in build-env.sh and the build tools required to build the kernel.

Since we clean up after ourselves in build-env.sh there's no reason to
make a temporary build container for this and then rip the entire rootfs
out for the actual kpv container. In addition this prevents accidentally
breaking apt with intermediary state files and clobbering the build
cache.

Signed-off-by: Randolph Sapp <rs@ti.com>