kernel_patch_verify: test_patch: Handle multiple files
When we have multiple files, we end up with a \n separated list. This
will not work correctly. Instead switch the usage to store the filenames
which may potentially contain spaces in an array and then convert it
into a list of quoted strings when passing it for the specific test.

Reported-by: Michael Nemanov <michael.nemanov@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>