kernel_patch_verify / HACK: Modify test of device tree with Makefile mod for dtb-
There is a bunch of warnings that 0day was reporting, which was
tracked down to CONFIG_OF_ALL_DTBS so that dtb+overlay are applied to
create test dtbs which are then checked correctly by dtbs_check.
Unfortunately, enabling CONFIG_OF_ALL_DTBS will cause *all* arch dtbs to
be built. Alternatively, we hand modify in the tool to enable the "dtb-"
which conventionally we have used for the test overlays so that these
are flagged and caught.

And with that we will end up getting a bunch more warnings and proper
checks of DT overlays. But we can work with just our platforms enabled
rather than the entire world of dtbs being built without the overhead
of CONFIG_OF_ALL_DTBS.

Examples:
https://lore.kernel.org/all/CAL_JsqKDMecB3dD2oU2MJo-_8G66K-mVTfNKWUBuL-=_tWpjtw@mail.gmail.com/#t
https://lore.kernel.org/all/CAL_JsqLxHKdQujWdSkGe6_cUftdgNgJ2PyVZR41A_LJ241_D3g@mail.gmail.com/
and so on.

Reported-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>